Refactored the way that messages are passed.
This change is more or less a rewrite at the heart of the system: 

Before, two different kinds of events were raised from users toward services, and two different callbacks were used from services toward users.

After this change, one kind of event with subclasses is raised from users toward services, and another kind of event with subclasses is raised from users toward events. A lot of relationship-tracking code that was needed in the old model was abandoned, and now we just use paired symmetric events and listeners.

In the process, a number of TODOs were added for further cleanup, and a number of bugs were incidentally discovered and squashed. This change is overbroad and messy; sorry. Hopefully this change will be part of getting the code base to a state where smaller, more surgical changes are possible.