Switched hardcoded values to SQLite
CharacterOwnershipIndex and WorldIndex are now fed by SQLite DB files instead of the old hardcoded implementations. The hardcoded implementations are still around for convenience (might we want them while we're testing?)

Default DB files are checked in, and copy to output directory on build if they're newer. (If you're editing your DB files directly in the bin directory, maybe that's not a good idea?) We will likely want to revisit this decision once code is programmatically writing to the DB; right now there is no user interface to do so.

The values in the default DB files currently match the hardcoded values; these values are explicitly not sane, they are particular to me. Getting sane values in there is left as an exercise to a future dev.